what a monstrocity
most of the data is pulling through now, but the player details page is all kinds of jacked up when you start throwing multiple years at it. Also for some reason 2019 is not working for my page at all so thats cool. Had to do a lot of modifying to the data structures to combine data for ALL the years given so it is pretty wretched but its getting the job done for now. definitely need to clean ALL of this up real bad